Relocate the refactored auction_prices functions to proper modules
