+++ crates/autopilot/src/database/mod.rs
@@ -10,7 +10,6 @@ use {
 };
 
 mod auction;
-pub mod auction_prices;
 pub mod competition;
 pub mod ethflow_events;
 pub mod events;
