+++ crates/orderbook/src/database/auctions.rs
@@ -1,6 +1,9 @@
 use {
     crate::dto,
+    alloy::primitives::Address,
     anyhow::{Context, Result},
+    bigdecimal::BigDecimal,
+    std::collections::HashMap,
 };
 
 impl super::Postgres {
@@ -30,5 +33,19 @@ impl super::Postgres {
         database::auction::last_used_auction_id(&mut ex)
             .await
             .context("could not fetch last used auction_id")
+    }
+
+    pub async fn fetch_latest_prices(&self) -> Result<HashMap<Address, BigDecimal>> {
+        let _timer = super::Metrics::get()
+            .database_queries
+            .with_label_values(&["fetch_latest_prices"])
+            .start_timer();
+
+        let mut ex = self.pool.acquire().await?;
+        Ok(database::auction::fetch_latest_prices(&mut ex)
+            .await?
+            .into_iter()
+            .map(|auction_price| (Address::new(auction_price.token.0), auction_price.price))
+            .collect::<HashMap<_, _>>())
     }
 }
