+++ crates/orderbook/src/database/mod.rs
@@ -1,5 +1,4 @@
 pub mod app_data;
-pub mod auction_prices;
 pub mod auctions;
 pub mod debug_report;
 mod fee_policies;
