+++ crates/orderbook/src/database/orders.rs
@@ -457,13 +457,12 @@ impl Postgres {
             }),
             self.execute_instrumented("fetch_latest_token_price", async {
                 let mut ex = self.pool.acquire().await?;
-                Ok(database::auction_prices::fetch_latest_token_price(
-                    &mut ex,
-                    ByteArray(token.0.0),
+                Ok(
+                    database::auction::fetch_latest_token_price(&mut ex, ByteArray(token.0.0))
+                        .await
+                        .map_err(anyhow::Error::from)?
+                        .and_then(|price| big_decimal_to_u256(&price)),
                 )
-                .await
-                .map_err(anyhow::Error::from)?
-                .and_then(|price| big_decimal_to_u256(&price)))
             })
         )?;
